No har me 3 filer: Main, CorotBeam og BeamModels. Ryddig
